Replaced double with int

+++ src/main/java/me/mrCookieSlime/Slimefun/Objects/SlimefunItem/machines/electric/gps/GPSTransmitter.java
@@ -39,7 +39,7 @@ public abstract class GPSTransmitter extends SimpleSlimefunItem<BlockTicker> {
 		});
 	}
 	
-	public abstract double getMultiplier(int y);
+	public abstract int getMultiplier(int y);
 	public abstract int getEnergyConsumption();
 
 	@Override
